brosario - Added the first part of the Trivia Maze.
I have just added all of the txt files that I think that we will need so
far. They are all orginized in a IntelliJ project. The only things that
contain anything are the TriviaMaze file (just realized that is
mispelled) and the Maze file. Pseudocode is a lot harder then I thought
it would be. I need some practice.