I have added the Player Object PseudoCode
So far it is pretty bare bones. The Player Object just contains the position of the player in relation to the maze 2-d array. The position of the player will be abstracted and can only be accessed through setters, not directly to the fields. These setters are also directionally based and I think all movement and orientation should use this standard. I also included the hint counter within this Object. I think this is good idea as it is the Player that has and uses the hints, so conceptually it makes sense to me. Also I can see multiplayer functionality being implemented easier this way, if we want to go that route.