Changed one small part of the useHint Method to be consistent with the rest of the pseudocode
Just changed the "decrement myAvailable hints" to "myAvailableHints = myAvailableHints -1" to be consistent with the rest of the pseudocode. I am not sure how close to actual code I should be and if it needs to be more "readable". As of right now I will continue to do this unless there is an issue.