Tiny changes made to the Maze Object pseudocode
Added a comma to the myDoors Map to be consistent with other pseudocode. Also made the addPlayer Method call the Player Constructor instead of setPosition for the player. Right now the pseudocode for the Player Object gives the player an initial position when constructed which seems more in line with how the program should function.