Changed gitignore file for project.
For my project structure I keep all of the my templates and references within the project. Don't know if this is a good way to do it , but that's what I do. So I just added the Object_Layout_2.0.jpg file and the Software Requirements and Specifications (SRS) Document Initial Draft folder to the gitignore file so I do not commit these file to the repository.